+++ src/main/java/com/meoguri/linkocean/util/querydsl/JoinInfoBuilder.java
@@ -1,4 +1,4 @@
-package com.meoguri.linkocean.util;
+package com.meoguri.linkocean.util.querydsl;
 
 import java.util.ArrayList;
 import java.util.Arrays;
